app/requirements.txt
pandas==2.1.0
mysql-connector-python==8.1.0
